Initial rustberry code commit
This introduces rustberry a small kernel for the Raspberry Pi 2 board.
Currently only the Raspberry Pi 2 Model B is supported.
This commit provides is an early blinker example.
It pulses GPIO16 that may blink an LED.

The code is new so it is probably not very useful in the current stage.

See the README.md for further info.